Add "project" to EnkoreJSRuntimeContextOptions. Why?
Because the field "originatingPackage" in EnkoreJSRuntimeContext doesn't work correctly right now.

I could do some tricks with the ''this'' variable but I'd rather keep it simple.

In the future, the API will look like this:

function someFactory(contextOptions: EnkoreJSRuntimeContextOptions, ...) {
    // originatingPackage is ''injected'' here.
    const context = createContext(contextOptions)

    // information about the current project (top level project so to speak)
    context.currentProject;

    // information about the current package (references the current package)
    context.currentPackage;

    context.log("hello")
}

So we are no longer passing down contexts, but rather options for a context.

I think this makes the most sense.

NB: the next few commits will adjust EnkoreJSRuntimeContext and EnkoreJSRuntimeContextOptions accordingly.