Some autogenerated files may depend on files inside project/build.
However, at the moment this is not possible because auto-generated files are created before the pre-processing stage has started.

If this flag is set to 'true', the file will be created _after_ pre-processing.

These files will be called "late autogenerated files" internally.

NB: I have yet to decide on whether late auto-generated files will be "pre-processed" or left as is.